+++ Internal/Debug App/Sources/Network/Networking+Models.swift
@@ -19,18 +19,12 @@ extension Networking {
         countryCode: CountryCode?,
         testParams: Test.Params?
     ) -> ClientSessionRequestBody {
-        
-        var metadataTestCaseDict: [String : Any]? = nil
-        if let metadataTestCaseStringValue = metadataTestCase {
-            metadataTestCaseDict = ["TEST_CASE": metadataTestCaseStringValue]
-        }
-        
         return ClientSessionRequestBody(
             customerId: customerId,
             orderId: "ios_order_id_\(String.randomString(length: 8))",
             currencyCode: currency,
             amount: nil,
-            metadata: metadataTestCaseDict,
+            metadata: nil,
             customer: ClientSessionRequestBody.Customer(
                 firstName: "John",
                 lastName: "Smith",
